added Makefile and systemd unit file
lod-api is can not be installed as local user with the provided
Makefile by calling `make` as root. This will furthermore
install a systemd template for the defined user in the Makefile to
manage lod-api as service.

To change the user under which the lod-api should run, overwrite the
envirnment variable `$User` as follows: `User=username make`

To start or enable the service, do:
`systemctl start/enable lod-api@username.service`