added resource's docCount to /topicsearch
if /topicsearch is used to query topics related to
a given query string this topic are rechecked how often
they are used within the resource index. This way one can
identify unimportened topics (i.e. docCount==0) and ignore
them later on in the process.

For now all topics are returned. Even those not mentioned in
the resources documents.