restructured size of returned topicsearch results
topics are evaluated in the score order returned from elasticsearch. The
process of gathering topics end if `size` valid topics are found. Valid
means that there are at least one document to this specific topic. If
there is not at least one document with this topic, this topic counts as
invalid. The number of invalid topics is also limited by `size`.